+++ README.md
@@ -144,10 +144,11 @@ MirroredObjectDefinition.create(SpaceFruit.class)
 
 
 ## Test support
-Ymer includes three test base classes (located in the `ymer-test` module) which can be used to verify that the defined `MongoConverter` can convert all mirrored space objects to bson, to test that data-migrations are applied as intended and also to check for @SpaceClass annotated classes that are not persisted.
+Ymer includes three test base classes which can be used to verify that the defined `MongoConverter` can convert all mirrored space objects to bson, to test that data-migrations are applied as intended and also to check for `@SpaceClass` annotated classes that are not persisted.
 
-These tests are implemented in both `junit4` and `junit5` using optional dependencies.
-`junit4` test cases are located in `com.avanza.junit4` while `junit5` test cases are located in `com.avanza.junit5`.
+These tests are available for `junit4` in the `ymer-test-junit4` module and for `junit5` in the `ymer-test-junit5`-module.
+The `ymer-test` module is available as an alias for `ymer-test-junit4` for backwards-compatibility and should not be depended on.
+
 The examples provided below are using `junit5`.
 
 ```java
@@ -230,16 +231,14 @@ Ymer is packed as a single jar file. Maven users can get Ymer using the followin
 </dependency>
 ```
 
-The test support is packed in a distinct jar using the following coordinates:
+The `junit5` test support is packed in a distinct jar using the following coordinates:
 ```xml
 <dependency>
   <groupId>com.avanza.ymer</groupId>
-  <artifactId>ymer-test</artifactId>
+  <artifactId>ymer-test-junit5</artifactId>
   <version>2.0.0</version>
 </dependency>
 ```
-
-In order to run the tests, `junit4` or `junit5` are also needed.
 
 ## License
 The Ymer Framework is released under version 2.0 of the [Apache License](https://www.apache.org/licenses/LICENSE-2.0).
