+++ ymer-test-junit4/.gitignore
@@ -1,0 +1,1 @@
+/target/
