+++ ymer-test-junit4/src/main/java/com/avanza/ymer/TestSpaceObject.java
