+++ ymer-test-junit4/src/test/resources/log4j2.xml
