+++ ymer-test-junit5/src/main/java/com/avanza/ymer/MirroredObjectTestHelper.java
