+++ ymer-test-junit5/src/test/resources/log4j2.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Configuration status="WARN" monitorInterval="60" shutdownHook="disable">
+<Appenders>
+    <Console name="Console" target="SYSTEM_OUT">
+        <PatternLayout pattern="%-4r [%t] %-5p %c %x - %m%n" />
+    </Console>
+</Appenders>
+<Loggers>
+    <Logger name="com.avanza" level="info" additivity="false">
+        <AppenderRef ref="Console" />
+    </Logger>
+    <Root level="warn">
+        <AppenderRef ref="Console" />
+    </Root>
+</Loggers>
+</Configuration>
