--- ymer-test/src/main/java/com/avanza/ymer/YmerMigrationTestBase.java
@@ -1,27 +1,0 @@
-/*
- * Copyright 2015 Avanza Bank AB
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.avanza.ymer;
-
-/**
- * @deprecated use {@link com.avanza.ymer.junit4.YmerMigrationTestBase} for JUnit 4
- */
-@Deprecated
-public abstract class YmerMigrationTestBase extends com.avanza.ymer.junit4.YmerMigrationTestBase {
-
-	public YmerMigrationTestBase(MigrationTest testCase) {
-		super(testCase);
-	}
-}
